requirements.txt
pypdf2
pyyaml
pynput
time
os
datetime
shutil
glob
halo
openai
textwrap
reportlab




